requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.5.2
attrs==25.3.0
blinker==1.8.2
certifi==2025.11.12
click==8.1.8
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.1
flasgger==0.9.7.1
flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
flask-mail==0.10.0
Flask-Migrate==4.1.0
flask-sqlalchemy==3.1.1
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.9.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mako==1.3.10
MarkupSafe==2.1.5
mistune==3.2.0
openai==2.2.0
packaging==25.0
pkgutil-resolve-name==1.3.10
pydantic==2.10.6
pydantic-core==2.27.2
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.3
referencing==0.35.1
rpds-py==0.20.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.44
tqdm==4.67.1
typing-extensions==4.13.2
werkzeug==3.0.6
zipp==3.20.2
